Rename component and remove code comment
