+++ src/components/ratingBolts.tsx
